+++ wall.cpp
@@ -2,10 +2,10 @@
 
 void Wall::draw(Tmpl8::Surface* surface, const vec2<float>& offset) {
 	surface->Bar(
-		static_cast<int>(pos_.x + offset.x), 
-		static_cast<int>(pos_.y + offset.y),
-		static_cast<int>(pos_.x + offset.x + size_.x),
-		static_cast<int>(pos_.y + offset.y + size_.y),
+		pos_.x + offset.x, 
+		pos_.y + offset.y,
+		pos_.x + offset.x + size_.x,
+		pos_.y + offset.y + size_.y,
 		0xffffff
 	);
 }
